requirements.txt
Django>=5.0,<5.1
psycopg[binary]>=3.1,<4
django-environ>=0.11,<1
djangorestframework>=3.15,<4
django-filter>=24.1,<25
django-jazzmin>=3.0,<4
gunicorn>=21.2,<22
whitenoise>=6.6,<7
pandas>=2.1,<3
openpyxl>=3.1,<4
django-cors-headers>=4.3,<5